+++ src/confluent_kafka/schema_registry/common/avro.py
@@ -25,6 +25,7 @@ __all__ = [
     'get_type',
     '_disjoint',
     '_resolve_union',
+    '_collapse_schema',
     'get_inline_tags',
     '_get_inline_tags_recursively',
     '_implied_namespace',
@@ -229,11 +230,64 @@ def _resolve_union(schema: AvroSchema, message: AvroMessage) -> Tuple[Optional[A
                     if dict_schema["name"] == dict_message['-type']:
                         return (dict_schema, dict_message)
             else:
-                validate(message, subschema)
+                validate(message, _collapse_schema(deepcopy(subschema)))
                 return (subschema, message)
         except:  # noqa: E722
             continue
     return (None, message)
+
+
+def _collapse_schema(schema: AvroSchema, encountered_references=None) -> AvroSchema:
+    """
+    Collapses a schema to conform to the Avro specification if it has been previously expanded.
+    Recursively replaces record, fixed, or enum definitions with their name when they have been already defined.
+    Mutates the incoming schema.
+
+    Args:
+        schema: An (expanded) Avro schema
+        encountered_references: A list of encountered references (used in the recursion)
+
+    Returns:
+        AvroSchema: A collapsed Avro schema.
+    """
+    if encountered_references is None:
+        encountered_references = []
+    if isinstance(schema, str):
+        return schema
+    elif isinstance(schema, list):
+        return [_collapse_schema(subschema, encountered_references) for subschema in schema]
+    elif isinstance(schema, dict):
+        schema_type = schema.get("type")
+        if schema_type == 'array':
+            schema["items"] = _collapse_schema(schema['items'], encountered_references)
+            return schema
+        elif schema_type == 'map':
+            schema["values"] = _collapse_schema(schema["values"], encountered_references)
+            return schema
+        elif schema_type == "record":
+            name = schema.get("name")
+            namespace = schema.get("namespace")
+            full_name = name if namespace is None or (name is not None and '.' in name) else f"{namespace}.{name}"
+            if full_name in encountered_references:
+                return schema["name"]
+            encountered_references.append(full_name)
+            if schema.get("aliases") is not None:
+                for alias in schema["aliases"]:
+                    full_alias = alias if namespace is None or '.' in alias else f"{namespace}.{alias}"
+                    encountered_references.append(full_alias)
+            schema["fields"] = _collapse_schema(schema["fields"], encountered_references)
+            return schema
+        elif schema_type == "fixed" or schema_type == "enum":
+            if schema["name"] in encountered_references:
+                return schema["name"]
+            encountered_references.append(schema["name"])
+            if schema.get("aliases") is not None:
+                for alias in schema["aliases"]:
+                    encountered_references.append(alias)
+            return schema
+        schema["type"] = _collapse_schema(schema["type"], encountered_references)
+        return schema
+    return schema
 
 
 def get_inline_tags(schema: AvroSchema) -> Dict[str, Set[str]]: